move pending tx update to tx history and also update qp packages
