+++ packages/arb-token-bridge-ui/package.json
@@ -31,7 +31,7 @@
     "graphql": "^16.8.1",
     "lodash-es": "^4.17.21",
     "next": "^13.5.6",
-    "next-query-params": "^4.2.3",
+    "next-query-params": "^5.0.0",
     "overmind": "^28.0.1",
     "overmind-react": "^29.0.1",
     "posthog-js": "^1.57.2",
@@ -45,7 +45,7 @@
     "sharp": "^0.32.6",
     "swr": "^2.1.2",
     "tailwind-merge": "^2.0.0",
-    "use-query-params": "^2.1.1",
+    "use-query-params": "^2.2.1",
     "wagmi": "^0.12.18",
     "zod": "^3.22.4",
     "zustand": "^4.3.9"
